NewInterface Push
UI was build but need binding with the properties. Under the UserInterface File you could do a solo run for the New UI. All Button event was settled.  Need binding with the back end event. There are limited  commit in the fxml files.   Press the "-" "+" mark in the IDE  for code structure to help you understand the interface structure.